feat: BettingResult 로직을 MatchResult 로직과 합침
